fix(home page): fix slider
