+++ src/components/filmCard/filmCard.hbs
@@ -1,4 +1,4 @@
-<a href={{href}} class="film-card" id={{id}}>
+<a class="film-card" id={{id}}>
     <article>
         <div class="film-card__image-container">
             <img src="{{image}}" alt={{title}} class="film-card__image">
