+++ src/components/topFilm/topFilm.hbs
@@ -1,4 +1,4 @@
-<a href="#" class="top-film">
+<a class="top-film">
     <section>
         <div class="top-film__content">
             <h1 class="top-film__title">{{title}}</h1>
