+++ src/helpers/ratingTypeHelper.js
@@ -1,0 +1,9 @@
+export const getRatingType = rating => {
+	if (rating >= 8) {
+		return 'high'
+	}
+	if (rating >= 5) {
+		return 'medium'
+	}
+	return 'low'
+}
