+++ src/mocks/films.js
@@ -1,36 +1,3 @@
-export const filmsMock = [
-	{
-		id: 1,
-		name: 'Назад в будущее',
-		category: 'fantasy',
-		year_of_publish: 2099,
-	},
-	{
-		id: 2,
-		name: 'Звёздные войны',
-		category: 'sci-fi',
-		year_of_publish: 1977,
-	},
-	{
-		id: 3,
-		name: 'Терминатор',
-		category: 'action',
-		year_of_publish: 1984,
-	},
-	{
-		id: 4,
-		name: 'Матрица',
-		category: 'sci-fi',
-		year_of_publish: 1999,
-	},
-	{
-		id: 5,
-		name: 'Властелин колец: Братство кольца',
-		category: 'fantasy',
-		year_of_publish: 2001,
-	},
-]
-
 export const TOPFILM = {
 	id: 2,
 	image: '/src/assets/img/dune.webp',
@@ -167,128 +134,5 @@ export const FILMS = [
 		duration: 113,
 		createdAt: new Date(Date.now() - 30 * 24 * 60 * 60 * 1000),
 		updatedAt: new Date(),
-	},
-]
-
-export const GENRES = [
-	{
-		id: 'a1b2c3d4-e5f6-7890-abcd-ef1234567890',
-		title: 'Аниме',
-		image: '/src/assets/img/genres/Аниме.png',
-	},
-	{
-		id: 'b2c3d4e5-f6g7-8901-bcde-f23456789012',
-		title: 'Биографии',
-		image: '/src/assets/img/genres/Биографии.png',
-	},
-	{
-		id: 'c3d4e5f6-g7h8-9012-cdef-345678901234',
-		title: 'Боевики',
-		image: '/src/assets/img/genres/Боевики.png',
-	},
-	{
-		id: 'd4e5f6g7-h8i9-0123-defg-456789012345',
-		title: 'Вестерны',
-		image: '/src/assets/img/genres/Вестерны.png',
-	},
-	{
-		id: 'e5f6g7h8-i9j0-1234-efgh-567890123456',
-		title: 'Детективы',
-		image: '/src/assets/img/genres/Детективы.png',
-	},
-	{
-		id: 'f6g7h8i9-j0k1-2345-fghi-678901234567',
-		title: 'Дорамы',
-		image: '/src/assets/img/genres/Дорамы.png',
-	},
-	{
-		id: 'g7h8i9j0-k1l2-3456-ghij-789012345678',
-		title: 'Драмы',
-		image: '/src/assets/img/genres/Драмы.png',
-	},
-	{
-		id: 'h8i9j0k1-l2m3-4567-hijk-890123456789',
-		title: 'Документальные',
-		image: '/src/assets/img/genres/Документальные.png',
-	},
-	{
-		id: 'i9j0k1l2-m3n4-5678-ijkl-901234567890',
-		title: 'Исторические',
-		image: '/src/assets/img/genres/Исторические.png',
-	},
-	{
-		id: 'j0k1l2m3-n4o5-6789-jklm-012345678901',
-		title: 'Комедии',
-		image: '/src/assets/img/genres/Комедии.png',
-	},
-	{
-		id: 'k1l2m3n4-o5p6-7890-klmn-123456789012',
-		title: 'Короткометражки',
-		image: '/src/assets/img/genres/Короткометражки.png',
-	},
-	{
-		id: 'l2m3n4o5-p6q7-8901-lmno-234567890123',
-		title: 'Криминал',
-		image: '/src/assets/img/genres/Криминальные.png',
-	},
-	{
-		id: 'm3n4o5p6-q7r8-9012-mnop-345678901234',
-		title: 'Мелодрамы',
-		image: '/src/assets/img/genres/Мелодрамы.png',
-	},
-	{
-		id: 'n4o5p6q7-r8s9-0123-nopq-456789012345',
-		title: 'Мистика',
-		image: '/src/assets/img/genres/Мистика.png',
-	},
-	{
-		id: 'o5p6q7r8-s9t0-1234-opqr-567890123456',
-		title: 'Музыкальные',
-		image: '/src/assets/img/genres/Музыкальные.png',
-	},
-	{
-		id: 'p6q7r8s9-t0u1-2345-pqrs-678901234567',
-		title: 'Мультфильмы',
-		image: '/src/assets/img/genres/Мультфильмы.png',
-	},
-	{
-		id: 'q7r8s9t0-u1v2-3456-qrst-789012345678',
-		title: 'Приключения',
-		image: '/src/assets/img/genres/Приключения.png',
-	},
-	{
-		id: 'r8s9t0u1-v2w3-4567-rstu-890123456789',
-		title: 'Ромком',
-		image: '/src/assets/img/genres/Ромкомы.png',
-	},
-	{
-		id: 's9t0u1v2-w3x4-5678-stuv-901234567890',
-		title: 'Семейные',
-		image: '/src/assets/img/genres/Семейные.png',
-	},
-	{
-		id: 't0u1v2w3-x4y5-6789-tuvw-012345678901',
-		title: 'Спортивные',
-		image: '/src/assets/img/genres/Спортивные.png',
-	},
-	{
-		id: 'u1v2w3x4-y5z6-7890-uvwx-123456789012',
-		title: 'Триллеры',
-		image: '/src/assets/img/genres/Триллеры.png',
-	},
-	{
-		id: 'v2w3x4y5-z6a7-8901-vwxy-234567890123',
-		title: 'Ужасы',
-		image: '/src/assets/img/genres/Ужасы.png',
-	},
-	{
-		id: 'w3x4y5z6-a7b8-9012-wxyz-345678901234',
-		title: 'Фантастика',
-		image: '/src/assets/img/genres/Фантастика.png',
-	},
-	{
-		id: 'x4y5z6a7-b8c9-0123-xyz-456789012345',
-		title: 'Фэнтези',
-		image: '/src/assets/img/genres/Фэнтези.png',
 	},
 ]
